Added routes for handling differernt requests , did app.js clean
