+++ package.json
@@ -14,6 +14,7 @@
     "ejs": "^2.6.1",
     "express": "^4.16.3",
     "lodash": "^4.17.20",
-    "mongoose": "^5.11.18"
+    "mongoose": "^5.11.18",
+    "path": "^0.12.7"
   }
 }
