+++ routes/Report.js
@@ -1,0 +1,10 @@
+const express = require('express');
+const router = express.Router();
+
+
+
+router.get("/report" , (req , res)=>{
+    res.render("compose");
+});
+
+module.exports = router;
